requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.7.2
    # via django
colorama==0.4.6
    # via qrcode
django==4.2.3
    # via -r src/requirements/requirements.in
pillow==10.0.0
    # via -r src/requirements/requirements.in
pypng==0.20220715.0
    # via qrcode
qrcode==7.4.2
    # via -r src/requirements/requirements.in
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via qrcode
tzdata==2023.3
    # via django
